Resolved conflicts and updated tests
