+++ model/event.go
@@ -1,4 +1,4 @@
-package main
+package model
 
 type Event struct {
 	Description string
