+++ model/interaction.go
@@ -1,4 +1,4 @@
-package main
+package model
 
 type Interaction struct {
 	ItemName string
