+++ model/models.go
@@ -4,15 +4,3 @@ type Describable interface {
 	SetDescription(description string)
 	GetDescription() string
 }
-
-type Event struct {
-	Description string
-	Outcome string
-	Triggered bool
-}
-
-type Interaction struct {
-	ItemName string
-	EntityName string
-	Event *Event
-}
